scripts/requirements.txt
niquests>=3.10.0
PyPDF2>=3.0.0
python-pptx>=1.0.0
python-docx>=1.0.0
rich>=13.0.0
playwright>=1.40.0
requests>=2.31.0
pyyaml>=6.0
